Ajuste QAS: Solucion de incidencias en usuarios finales de la sede San Borja

+++ src/rp3/auna/sync/Agente.java
@@ -138,20 +138,29 @@ public class Agente {
 			try {
 				webService.invokeWebService();
 				JSONObject jObject = webService.getJSONObjectResponse();
-                Log.d(TAG,"getAgente:"+jObject.toString());
-				PreferenceManager.setValue(Contants.KEY_IDAGENTE, jObject.getInt(Contants.KEY_IDAGENTE));
-				if(!jObject.isNull(Contants.KEY_IDRUTA))
-                    PreferenceManager.setValue(Contants.KEY_IDRUTA, jObject.getInt(Contants.KEY_IDRUTA));
-                if(!jObject.isNull(Contants.KEY_FOTO))
-                    PreferenceManager.setValue(Contants.KEY_FOTO, jObject.getString(Contants.KEY_FOTO));
-				PreferenceManager.setValue(Contants.KEY_ES_SUPERVISOR, jObject.getBoolean(Contants.KEY_ES_SUPERVISOR));
-                if(!jObject.isNull(Contants.KEY_ID_SUPERVISOR))
-                    PreferenceManager.setValue(Contants.KEY_ID_SUPERVISOR, jObject.getInt(Contants.KEY_ID_SUPERVISOR));
-				PreferenceManager.setValue(Contants.KEY_ES_AGENTE, jObject.getBoolean(Contants.KEY_ES_AGENTE));
-				PreferenceManager.setValue(Contants.KEY_ES_ADMINISTRADOR, jObject.getBoolean(Contants.KEY_ES_ADMINISTRADOR));
-				PreferenceManager.setValue(Contants.KEY_CARGO, jObject.getString(Contants.KEY_CARGO));
-                if(!jObject.isNull(Contants.KEY_DESCUENTO_MAXIMO))
-                    PreferenceManager.setValue(Contants.KEY_DESCUENTO_MAXIMO, jObject.getInt(Contants.KEY_DESCUENTO_MAXIMO));
+				if(jObject!=null){
+                    Log.d(TAG,"getAgente:"+jObject.toString());
+                    if(!jObject.isNull(Contants.KEY_IDAGENTE)){
+                        PreferenceManager.setValue(Contants.KEY_IDAGENTE, jObject.getInt(Contants.KEY_IDAGENTE));
+                    }else{
+                        return SyncAdapter.SYNC_EVENT_AUTH_ERROR;
+                    }
+
+                    if(!jObject.isNull(Contants.KEY_IDRUTA))
+                        PreferenceManager.setValue(Contants.KEY_IDRUTA, jObject.getInt(Contants.KEY_IDRUTA));
+                    if(!jObject.isNull(Contants.KEY_FOTO))
+                        PreferenceManager.setValue(Contants.KEY_FOTO, jObject.getString(Contants.KEY_FOTO));
+                    PreferenceManager.setValue(Contants.KEY_ES_SUPERVISOR, jObject.getBoolean(Contants.KEY_ES_SUPERVISOR));
+                    if(!jObject.isNull(Contants.KEY_ID_SUPERVISOR))
+                        PreferenceManager.setValue(Contants.KEY_ID_SUPERVISOR, jObject.getInt(Contants.KEY_ID_SUPERVISOR));
+                    PreferenceManager.setValue(Contants.KEY_ES_AGENTE, jObject.getBoolean(Contants.KEY_ES_AGENTE));
+                    PreferenceManager.setValue(Contants.KEY_ES_ADMINISTRADOR, jObject.getBoolean(Contants.KEY_ES_ADMINISTRADOR));
+                    PreferenceManager.setValue(Contants.KEY_CARGO, jObject.getString(Contants.KEY_CARGO));
+                    if(!jObject.isNull(Contants.KEY_DESCUENTO_MAXIMO))
+                        PreferenceManager.setValue(Contants.KEY_DESCUENTO_MAXIMO, jObject.getInt(Contants.KEY_DESCUENTO_MAXIMO));
+                }else{
+				    return SyncAdapter.SYNC_EVENT_AUTH_ERROR;
+                }
 			} catch (HttpResponseException e) {
 				if(e.getStatusCode() == HttpConnection.HTTP_STATUS_UNAUTHORIZED)
 					return SyncAdapter.SYNC_EVENT_AUTH_ERROR;
@@ -344,6 +353,9 @@ public class Agente {
 			try {
 				webService.invokeWebService();
 				JSONObject jObject = webService.getJSONObjectResponse();
+				if(jObject.isNull(Contants.KEY_ALARMA_INTERVALO)){
+                    return SyncAdapter.SYNC_EVENT_AUTH_ERROR;
+                }
 				PreferenceManager.setValue(Contants.KEY_ALARMA_INICIO, Convert.getDateFromDotNetTicks(jObject.getLong(Contants.KEY_ALARMA_INICIO)));
 				PreferenceManager.setValue(Contants.KEY_ALARMA_FIN, Convert.getDateFromDotNetTicks(jObject.getLong(Contants.KEY_ALARMA_FIN)));
 				PreferenceManager.setValue(Contants.KEY_ALARMA_INTERVALO, jObject.getInt(Contants.KEY_ALARMA_INTERVALO));
@@ -390,6 +402,9 @@ public class Agente {
             }
 
             JSONArray types = webService.getJSONArrayResponse();
+            if(types==null){
+                return SyncAdapter.SYNC_EVENT_AUTH_ERROR;
+            }
             Log.d(TAG,"getAgenteS:"+types.toString());
             rp3.auna.models.Agente.deleteAll(db, Contract.Agente.TABLE_NAME);
